Fix debt share snapshot id type (#1023)
* fix debt share snapshot id

* fix debt share period id to be saved as uint128

+++ packages/synthetix-governance/contracts/interfaces/IDebtShare.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.0;
 
 interface IDebtShare {
-    function balanceOfOnPeriod(address account, uint128 periodId) external view returns (uint);
+    function balanceOfOnPeriod(address account, uint periodId) external view returns (uint);
 }
